requirements.txt
anyio==3.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
dill==0.3.6
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
keyboard==0.13.5
llvmlite==0.40.1rc1
lxml==4.9.2
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
multiprocess==0.70.14
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
pathos==0.3.0
pox==0.3.2
ppft==1.7.6.6
pycparser==2.21
PyExecJS==1.5.1
regex==2023.6.3
requests==2.31.0
scipy==1.10.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
translators==5.7.6
typing_extensions==4.6.3
urllib3==2.0.3
voicevox-client==0.2.0
zipp==3.15.0
